requirements.txt
xmltodict
xlsxwriter
numpy
pandas
plotly
opencv-python
scipy
seaborn
flask
psutil
requests
Flask-SQLAlchemy
flask-login
Flask-WTF
email_validator
gunicorn
google-cloud-storage
psycopg2-binary
SQLAlchemy[postgresql]
kaleido